corrected logging of high port numbers

+++ HttpServer.cpp
@@ -100,12 +100,10 @@ bool HttpServer::init ( int16_t port,
 		m_ssltcp.reset();
 	}
 	// log an innocent msg
-	log(LOG_INIT,"http: Listening on TCP port %i with sd=%i", 
-	    port, m_tcp.m_sock );
+	log(LOG_INIT,"http: Listening on TCP port %" PRIu16 " with sd=%i", (uint16_t)port, m_tcp.m_sock );
 	// log for https
 	if (m_ssltcp.m_ready)
-		log(LOG_INIT,"https: Listening on TCP port %i with sd=%i", 
-	    	    sslPort, m_ssltcp.m_sock );
+		log(LOG_INIT,"https: Listening on TCP port %" PRIu16 " with sd=%i", (uint16_t)sslPort, m_ssltcp.m_sock );
 
 	return true;
 }
